fix issue with Either deserialization with case classes
EitherDeserializer was reusing a JsonParser first when it attempted to
parse the Left, then if that failed and it tried to parse the Right. If
the Left and Right sides look similar enough that the nextToken method
was ever called on the parser, then the parsing of the Right could fail.